Create collatz
When the number is odd, the next step is to multiply it by 3 and then add 1.  When the number is even, divide the number by two.  It always seems to reach 1, however it cannot (yet) be proved that it always will.  Whenever the number reaches a power of two (2^x) then the game will end.  Interesting to see how high the numbers can get.